test: reorganize timing specs by concern
